rename relationalmodels to relations, adapter to adapters
