--- lib/grape/roar/extensions/relational_models/validations.rb
@@ -1,3 +1,0 @@
-require 'grape/roar/extensions/relational_models/validations/misc'
-require 'grape/roar/extensions/relational_models/validations/active_record'
-require 'grape/roar/extensions/relational_models/validations/mongoid'
