+++ lib/grape/roar/extensions/relations/adapters/active_record.rb
@@ -1,8 +1,8 @@
 module Grape
   module Roar
     module Extensions
-      module RelationalModels
-        module Adapter
+      module Relations
+        module Adapters
           class ActiveRecord < Base
             include Validations::ActiveRecord
