+++ lib/grape/roar/extensions/relations/validations/mongoid.rb
@@ -1,7 +1,7 @@
 module Grape
   module Roar
     module Extensions
-      module RelationalModels
+      module Relations
         module Validations
           module Mongoid
             include Validations::Misc
