+++ spec/extensions/relations/adapters/active_record_spec.rb
@@ -1,12 +1,12 @@
 require 'spec_helper'
-require 'mongoid'
+require 'active_record'
 
-describe Grape::Roar::Extensions::RelationalModels::Adapter::Mongoid do
-  let(:model) { Class.new.tap { |c| c.include(::Mongoid::Document) } }
+describe Grape::Roar::Extensions::Relations::Adapters::ActiveRecord do
+  let(:model) { Class.new(ActiveRecord::Base) }
   subject { described_class.new(model) }
 
   context '.valid_for?' do
-    it 'is only valid for classes that mixed in Mongoid::Document' do
+    it 'is only valid for ActiveRecord::Base descendants' do
       expect(described_class.valid_for?(model)).to eql(true)
       expect(described_class.valid_for?('foo')).to eql(false)
     end
@@ -15,14 +15,14 @@ describe Grape::Roar::Extensions::RelationalModels::Adapter::Mongoid do
   context '#collection_methods' do
     it 'should return all collection methods' do
       expect(subject.collection_methods)
-        .to match_array(%i(embeds_many has_many has_and_belongs_to_many))
+        .to match_array(%i(has_many has_and_belongs_to_many))
     end
   end
 
   context '#single_entity_methods' do
     it 'should return all single entity methods' do
       expect(subject.single_entity_methods)
-        .to match_array(%i(has_one belongs_to embeds_one))
+        .to match_array(%i(has_one belongs_to))
     end
   end
 end
